Move SQL-related functions and constants into `db.utils` module
Move a utility method and some constants into the `db.utils` module
to cleanup `datamodel`.
